chore: 불필요한 Extension 파일 제거 및 코드 정렬
