+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/FirstColumn/SellPrice/SellPriceViewCell.swift
@@ -17,16 +17,11 @@ final class SellPriceViewCell: Cell {
     }
     
     override init(frame: CGRect) {
-
+        
         super.init(frame: frame)
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
     }
-    
-//    override func prepareForReuse() {
-//        self.sellPriceLabel.text = nil
-//        self.sellPriceRateLabel.text = nil
-//    }
 }
