+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/FirstColumn/SellQuantity/SellQuantityViewCell.swift
@@ -14,12 +14,8 @@ final class SellQuantityViewCell: Cell {
     override init(frame: CGRect) {
         super.init(frame: frame)
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
     }
-    
-//    override func prepareForReuse() {
-//        self.sellQuantityLabel.text = nil
-//    }
 }
