+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/FirstColumn/TradeDescription/TopViewCell.swift
@@ -15,13 +15,8 @@ final class TopViewCell: Cell {
     override init(frame: CGRect) {
         super.init(frame: frame)
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
     }
-    
-//    override func prepareForReuse() {
-//        self.tradeVolumeLabel.text = nil
-//        self.tradeValueLabel.text = nil
-//    }
 }
