+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/SecondColumn/BuyPrice/BuyPriceViewCell.swift
@@ -15,13 +15,8 @@ final class BuyPriceViewCell: Cell {
     override init(frame: CGRect) {
         super.init(frame: frame)
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
     }
- 
-//    override func prepareForReuse() {
-//        self.buyPriceLabel.text = nil
-//        self.buyPriceRateLabel.text = nil
-//    }
 }
