+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/SecondColumn/ConclusionTableView/ContentViewCell/ConcludedQuantity/ConcludedQuantityTableViewCell.swift
@@ -13,17 +13,14 @@ class ConcludedQuantityTableViewCell: UITableViewCell {
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
     }
-
+    
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
     }
     
     override func prepareForReuse() {
-//        self.concludedPriceLabel.text = nil
-//        self.concludedQuantityLabel.text = nil
+        self.concludedPriceLabel.text = nil
+        self.concludedQuantityLabel.text = nil
     }
 }
