+++ CryptocurrencyExchange/CryptocurrencyExchange/Orderbook/ContentView/SecondColumn/ConclusionTableView/ContentViewCell/FasteningStrength/FasteningStrengthTableViewCell.swift
@@ -12,12 +12,13 @@ class FasteningStrengthTableViewCell: UITableViewCell {
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
     }
     
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+    }
+    
+    override func prepareForReuse() {
+        self.fasteningStrengthLabel.text = nil
+    }
 }
